\name{cvm.test-package}
\alias{cvm.test-package}
\docType{package}
\title{
Discrete Cramer von Mises Goodness-of-Fit Test
}
\description{
Computes the test statistics for doing one sample Cramer von Mises Goodness-of-Fit Test
and calculates asymptotic p-values. Package tries to follow the same guidelines as ks.test in the R-base.
}
\details{
\tabular{ll}{
Package: \tab cvm.test\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-08-18\cr
License: \tab GPL (>= 2.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Taylor B. Arnold and John W. Emerson

Maintainer: Taylor B. Arnold <taylor.arnold@yale.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ Discrete Cramer von Mises }
